Add basic config loader built on himl
The goal is for all input parameters to the analysis scripts to be drawn from a hierarchic structure of config files. Thus, parameters and configurations can be specified at whatever level is most convenient, for example specifying a script input value that is the same for a large number of individual datasets at the level of a "experiment" or "collection".