Add basic records db interface with sqlite3
The records db is intended to store a ledger of all the experiments that were carried out. The data in this ledger must be sufficient to reconstruct everything necessary for carrying out the analysis of the corresponding dataset(s), ie reconstructing the config path and locating the data file itself.

Currently, the database attributes/formatting is hard-coded with four basic keys: experiment, meas_id, cooldown, and meas_type. In the near future this will be expanded to include at least a sort_datetime to enable locating the data file in the Labber data tree, as well as potentially other keys/columns.