Change ConfLoader directory logic
In order to simplify the user experience, the ConfLoader would previously append the configuration base directory automatically. While this is still desired behaviour for the end user, it was failing when the cwd was different to the desired base dir of the conf tree. After looking at the himl source, it turns out this appears to be an easy fix as leaving the `cwd` arg as None causes himl to use `os.getcwd()` and calculate the relative path from that, which is what we want. For the user, it should be easy to use `os.path.abspath()` to still use relative paths in terms of the code semantics even when an absolute path is required by the methods, as is now the case.